+++ pages/_app.js
@@ -5,7 +5,7 @@ export default function App({ Component, pageProps }) {
   return (
     <>
       <TopBar />
-      <Component {...pageProps} />;
+      <Component {...pageProps} />
     </>
 
   );
